[Edit]subjectId Long으로 수정

+++ app/src/main/java/com/example/allclear/schedule/Schedule.java
@@ -12,7 +12,7 @@ import java.io.Serializable;
 
 public class Schedule implements Serializable {
     @PrimaryKey(autoGenerate = true)
-    private int subjectId;
+    private Long subjectId;
     private String subjectName;
     private String professor;
     private int classDay;
@@ -23,9 +23,9 @@ public class Schedule implements Serializable {
     private String textColor= "#ffffff";
     public Long timetableId;  // 속한 시간표의 ID
 
-    public int getSubjectId() {return subjectId;}
+    public Long getSubjectId() {return subjectId;}
 
-    public void setSubjectId(int subjectId) {
+    public void setSubjectId(Long subjectId) {
         this.subjectId = subjectId;
     }
